reactDemo/server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-cors
python-dotenv
flask-session
redis
requests
gunicorn
# celery